+++ src/App.tsx
@@ -3,7 +3,7 @@ import { Routes, Route } from 'react-router-dom';
 
 import Layout from './pages/Layout';
 
-import Home from './pages/Home/Home';
+import Home from './pages/Home';
 import CampDetails from './pages/CampDetail/CampDetails';
 import CampApply from './pages/CampApply/CampApply';
 import Community from './pages/Community/Community';
